Min Energy Consumption
Cryptocurrency mining can consume significant amounts of energy, leading to environmental concerns. 
Improving the energy consumption of cryptocurrency mining can be achieved by implementing more energy-efficient algorithms and technologies. Here are a few code-based solutions that can help reduce the energy consumption of cryptocurrency mining. Switch to Proof-of-Stake (PoS) consensus algorithm: PoS is an alternative to the traditional Proof-of-Work (PoW) algorithm that is used by many cryptocurrencies, including Dogecoin. PoS uses significantly less energy as it does not require miners to solve complex mathematical problems in order to validate transactions. Here's an example code in C++ for switching to PoS